moved ejml tests to separate module
