+++ core/pom.xml
@@ -8,50 +8,22 @@
         <version>0.4-SNAPSHOT</version>
     </parent>
 
+    <name>Mat-File IO / Core</name>
     <artifactId>mat-file-io</artifactId>
     <packaging>jar</packaging>
 
     <properties>
-
-        <!-- stay compatible with MATLAB >2009a -->
-        <javaTarget>1.6</javaTarget>
-
+        <javaTarget>1.6</javaTarget>  <!-- compatible with MATLAB >2009a -->
         <javaModuleName>us.hebi.matlab.mat</javaModuleName>
-
     </properties>
 
-    <!-- Dependency management (keep in alphabetical order) -->
     <dependencies>
 
-        <dependency> <!-- for custom serialization tests -->
-            <groupId>org.ejml</groupId>
-            <artifactId>ejml-all</artifactId>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency> <!-- Unit Tests -->
+        <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
         </dependency>
 
     </dependencies>
-
-    <!-- Build Process -->
-    <build>
-
-        <!-- Include the README, NOTICE and LICENSE files: -->
-        <resources>
-            <resource>
-                <directory>${project.basedir}/..</directory>
-                <includes>
-                    <include>README*</include>
-                    <include>NOTICE*</include>
-                    <include>LICENSE*</include>
-                    <include>CHANGELOG*</include>
-                </includes>
-            </resource>
-        </resources>
-
-    </build>
 
 </project>
