+++ mat-ejml/src/main/java/us/hebi/matlab/mat/ejml/FMatrixSparseCSCWrapper.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,19 +18,19 @@
  * #L%
  */
 
-package us.hebi.matlab.mat.tests.serialization.ejml;
+package us.hebi.matlab.mat.ejml;
 
-import org.ejml.data.DMatrixSparseCSC;
+import org.ejml.data.FMatrixSparseCSC;
 import us.hebi.matlab.mat.format.Mat5;
 import us.hebi.matlab.mat.format.Mat5Serializable;
+import us.hebi.matlab.mat.format.Mat5Type;
+import us.hebi.matlab.mat.format.Mat5WriteUtil;
 import us.hebi.matlab.mat.types.AbstractArray;
 import us.hebi.matlab.mat.types.MatlabType;
 import us.hebi.matlab.mat.types.Sink;
 
 import java.io.IOException;
 
-import static us.hebi.matlab.mat.format.Mat5Type.Double;
-import static us.hebi.matlab.mat.format.Mat5Type.*;
 import static us.hebi.matlab.mat.format.Mat5WriteUtil.*;
 
 /**
@@ -42,15 +42,15 @@ import static us.hebi.matlab.mat.format.Mat5WriteUtil.*;
  *
  * @author Florian Enner
  */
-class DMatrixSparseCSCWrapper extends AbstractArray implements Mat5Serializable, Mat5Serializable.Mat5Attributes {
+class FMatrixSparseCSCWrapper extends AbstractArray implements Mat5Serializable, Mat5Serializable.Mat5Attributes {
 
     @Override
     public int getMat5Size(String name) {
         return Mat5.MATRIX_TAG_SIZE
                 + computeArrayHeaderSize(name, this)
-                + Int32.computeSerializedSize(getNumRowIndices())
-                + Int32.computeSerializedSize(getNumColIndices())
-                + Double.computeSerializedSize(getNzMax());
+                + Mat5Type.Int32.computeSerializedSize(getNumRowIndices())
+                + Mat5Type.Int32.computeSerializedSize(getNumColIndices())
+                + Mat5Type.Single.computeSerializedSize(getNzMax());
     }
 
     @Override
@@ -59,23 +59,23 @@ class DMatrixSparseCSCWrapper extends AbstractArray implements Mat5Serializable,
         writeArrayHeader(name, isGlobal, this, sink);
 
         // Row indices (MATLAB requires at least 1 entry)
-        Int32.writeTag(getNumRowIndices(), sink);
-        if (matrix.getNonZeroLength() == 0) {
+        Mat5Type.Int32.writeTag(getNumRowIndices(), sink);
+        if (sparse.getNonZeroLength() == 0) {
             sink.writeInt(0);
         } else {
-            sink.writeInts(matrix.nz_rows, 0, getNumRowIndices());
+            sink.writeInts(sparse.nz_rows, 0, getNumRowIndices());
         }
-        Int32.writePadding(getNumRowIndices(), sink);
+        Mat5Type.Int32.writePadding(getNumRowIndices(), sink);
 
         // Column indices
-        Int32.writeTag(getNumColIndices(), sink);
-        sink.writeInts(matrix.col_idx, 0, getNumColIndices());
-        Int32.writePadding(getNumColIndices(), sink);
+        Mat5Type.Int32.writeTag(getNumColIndices(), sink);
+        sink.writeInts(sparse.col_idx, 0, getNumColIndices());
+        Mat5Type.Int32.writePadding(getNumColIndices(), sink);
 
         // Non-zero values
-        Double.writeTag(getNzMax(), sink);
-        sink.writeDoubles(matrix.nz_values, 0, getNzMax());
-        Double.writePadding(getNzMax(), sink);
+        Mat5Type.Single.writeTag(getNzMax(), sink);
+        sink.writeFloats(sparse.nz_values, 0, getNzMax());
+        Mat5Type.Single.writePadding(getNzMax(), sink);
 
     }
 
@@ -85,16 +85,16 @@ class DMatrixSparseCSCWrapper extends AbstractArray implements Mat5Serializable,
     }
 
     private int getNumRowIndices() {
-        return Math.max(1, matrix.getNonZeroLength());
+        return Math.max(1, sparse.getNonZeroLength());
     }
 
     private int getNumColIndices() {
-        return matrix.getNumCols() + 1;
+        return sparse.getNumCols() + 1;
     }
 
     @Override
     public int getNzMax() {
-        return matrix.getNonZeroLength();
+        return sparse.getNonZeroLength();
     }
 
     @Override
@@ -109,33 +109,32 @@ class DMatrixSparseCSCWrapper extends AbstractArray implements Mat5Serializable,
 
     @Override
     public int[] getDimensions() {
-        dims[0] = matrix.getNumRows();
-        dims[1] = matrix.getNumCols();
+        dims[0] = sparse.getNumRows();
+        dims[1] = sparse.getNumCols();
         return dims;
     }
 
-    DMatrixSparseCSCWrapper(DMatrixSparseCSC matrix) {
-        super(Mat5.dims(matrix.getNumRows(), matrix.getNumCols()));
-        if (!matrix.indicesSorted)
+    public FMatrixSparseCSCWrapper(FMatrixSparseCSC sparse) {
+        super(Mat5.dims(sparse.getNumRows(), sparse.getNumCols()));
+        if (!sparse.indicesSorted)
             throw new IllegalArgumentException("Indices must be sorted!");
-        this.matrix = matrix;
+        this.sparse = sparse;
     }
 
     @Override
     public void close() throws IOException {
     }
 
-    final DMatrixSparseCSC matrix;
+    final FMatrixSparseCSC sparse;
 
     @Override
     protected int subHashCode() {
-        return matrix.hashCode();
+        return sparse.hashCode();
     }
 
     @Override
     protected boolean subEqualsGuaranteedSameClass(Object otherGuaranteedSameClass) {
-        DMatrixSparseCSCWrapper other = (DMatrixSparseCSCWrapper) otherGuaranteedSameClass;
-        return other.matrix.equals(matrix);
+        FMatrixSparseCSCWrapper other = (FMatrixSparseCSCWrapper) otherGuaranteedSameClass;
+        return other.sparse.equals(sparse);
     }
-
 }
