+++ pom.xml
@@ -65,6 +65,7 @@
 
     <modules>
         <module>core</module>
+        <module>mat-ejml</module>
     </modules>
 
     <properties>
